fixed more compiler warnings
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@4951 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ surface/include/pcl/surface/grid_projection.h
@@ -242,7 +242,7 @@ namespace pcl
       getCellIndex (const Eigen::Vector4f &p, Eigen::Vector3i& index) const
       {
         for (int i = 0; i < 3; ++i)
-          index[i] = (p[i] - min_p_(i))/leaf_size_;
+          index[i] = static_cast<int> ((p[i] - min_p_(i)) / leaf_size_);
       }
 
       /** \brief Given the 3d index (x, y, z) of the cell, get the 
@@ -254,7 +254,10 @@ namespace pcl
       getCellCenterFromIndex (const Eigen::Vector3i &index, Eigen::Vector4f &center) const
       {
         for (int i = 0; i < 3; ++i)
-          center[i] = min_p_[i] + index[i] * leaf_size_ + leaf_size_/2;
+          center[i] = 
+            min_p_[i] + static_cast<float> (index[i]) * 
+            static_cast<float> (leaf_size_) + 
+            static_cast<float> (leaf_size_) / 2.0f;
       }
 
       /** \brief Given cell center, caluate the coordinates of the eight vertices of the cell
